refact(struct): site structure and code syntax
