--- README.md~
@@ -1,13 +1,0 @@
-# NFL Queries
-
-##The best kind of stat is the one you regret finding out about.
-
-Run queries on an NFL players database (Updated on February 2015) to find out weird facts that will still be as interesting as ESPN's SportsCenter coverage of the NFL off-season. 
-
-The players.json file was extracted from the Python ['nflgame'](https://github.com/BurntSushi/nflgame) library (Which is totally awesome!)
-
-TODO:
-- Finish the project
-- Examples
-- Add teams stats support
-
