+++ index.html
@@ -1,126 +1,109 @@
 <!DOCTYPE html>
 <html lang="en">
-	<head>
-		<meta http-equiv="content-type" content="text/html; charset=UTF-8">
-		<meta charset="utf-8">
-		<title>NFL Queries</title>
-		<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
-		<link href="css/bootstrap.min.css" rel="stylesheet">
-		<link href="http://netdna.bootstrapcdn.com/font-awesome/3.2.1/css/font-awesome.min.css" rel="stylesheet">
-        <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.2.0/css/font-awesome.min.css" rel="stylesheet">
 
-		<!--[if lt IE 9]>
+<head>
+	<meta http-equiv="content-type" content="text/html; charset=UTF-8">
+	<meta charset="utf-8">
+	<title>NFL Queries</title>
+	<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
+	<link href="assets/css/bootstrap.min.css" rel="stylesheet">
+	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css">
+
+	<!--[if lt IE 9]>
 			<script src="//html5shim.googlecode.com/svn/trunk/html5.js"></script>
 		<![endif]-->
-		<link href="css/styles.css" rel="stylesheet">
-	</head>
-	<body>
+	<link href="assets/css/styles.css" rel="stylesheet">
+</head>
 
-<nav class="navbar navbar-default navbar-fixed-top">
-  <div class="container">
-      
-      <div class="navbar-header">
-      <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1">
-        <span class="sr-only">Toggle navigation</span>
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-      </button>
-      <a class="navbar-brand">NFL Queries</a>
-    </div>
-      
-      
-    <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-      
-      <ul class="nav navbar-nav navbar-right">
+<body>
 
-      </ul>
-        
-        <ul class="nav navbar-nav navbar-right nav-pills">
-                        
-            <li>
-            <a href="/nflqueries/" class="btn btn-social-icon" target="_blank">
+	<nav class="navbar navbar-default navbar-fixed-top">
+		<div class="container">
+			<div class="navbar-header">
+				<button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1">
+					<span class="sr-only">Toggle navigation</span>
+					<span class="icon-bar"></span>
+					<span class="icon-bar"></span>
+					<span class="icon-bar"></span>
+				</button>
+				<a class="navbar-brand">NFL Queries</a>
+			</div>
+
+
+			<div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+				<ul class="nav navbar-nav navbar-right">
+				</ul>
+				<ul class="nav navbar-nav navbar-right nav-pills">
+					<li>
+						<a href="/" class="btn btn-social-icon">
                 Home
             </a>
-                    </li>
-            
-            <li>
-            <a href="/nflqueries/about.html" class="btn btn-social-icon" target="_blank">
+					</li>
+					<li>
+						<a href="/about.html" class="btn btn-social-icon">
                 About
             </a>
-            </li> 
-            
-            </ul>
-        
-    </div><!-- /.navbar-collapse -->
-  </div>
-</nav>
-        
-<div class="container-main">
-      <div class="row">
-       <div class="col-lg-12" style="text-align:center;margin-top:1.5%"><h1 class="nflqueries">NFL Queries</h1>
-          <p class="lead">The best kind of stat is the one you regret finding out about.</p>
-          <p class="readable queries">Made with love by <a href="http://twitter.com/FanaHOVA" target="_blank">@FanaHOVA</a></p>
-          
-                        <p class="readable queries"><br>Example queries: <br>
-                  
-              Players with number 15 whose first name starts with G<br>
-              Players from Mississippi who wore number 10<br>
-              Packers players with number 12<br>
-              Players from UCLA that weigh more than 150lbs
-                </p></div>
-          
+					</li>
+				</ul>
+			</div>
+		</div>
+	</nav>
 
-          
-        <div class="col-lg-12 text-center v-center">
-            
-            <div id="noresults" class="noresults"></div>
-            <div id="nofresults" class="success"></div>
-            
-          <form class="col-lg-12" id="query-input">
-            <div class="querybar input-group">
-            <input class="form-control input-lg" id="query" placeholder="Ask me stuff" type="text">
-              <span class="input-group-btn"><button class="btn btn-lg btn-primary" type="button" onclick="getvalues()">Go!</button></span>
-            </div>
-          </form>
-        </div>
-        
-      </div> <!-- /row -->
-  	  <div class="row">
-        <div class="col-lg-12 text-center v-center" >
-            
-            <div id="result"></div>
-          
-        </div>
-            
-      </div>
-  
-</div> <!-- /container full -->
+	<div class="container-main">
+		<div class="row">
+			<div class="col-lg-12" style="text-align:center;margin-top:1.5%">
+				<h1 class="nflqueries">NFL Queries</h1>
+				<p class="lead">The best kind of stat is the one you regret finding out about.</p>
+				<p class="readable queries"><i class='fa fa-code'></i> with <i class='fa fa-heart'></i> by <a href="http://twitter.com/FanaHOVA" target="_blank">@FanaHOVA</a></p>
 
-	<!-- script references -->
+				<p class="readable queries">
+					<br>Example queries:
+					<br>Players with number 15 whose first name starts with G
+					<br>Players from Mississippi who wore number 10
+					<br>Packers players with number 12
+					<br>Players from UCLA that weigh more than 150lbs
+				</p>
+			</div>
 
-        <script src="http://ajax.googleapis.com/ajax/libs/jquery/2.0.2/jquery.min.js"></script>
-		<script src="js/bootstrap.min.js"></script>
-        <script src="js/regex.js"></script>
-        
-        <script type="text/javascript">
 
-    $('#query-input').submit(function(ev){
-    ev.preventDefault();
-    getvalues();
-    return false;
-});
 
-  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-  (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-  m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-  })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
+			<div class="col-lg-12 text-center v-center">
 
-  ga('create', 'UA-47296378-1', 'auto');
-  ga('send', 'pageview');
+				<div id="noresults" class="noresults"></div>
+				<div id="nofresults" class="success"></div>
 
-            
-</script>
-        
-	</body>
+				<form class="col-lg-12" id="query-input">
+					<div class="querybar input-group">
+						<input class="form-control input-lg" id="query" placeholder="Ask me stuff" type="text">
+						<span class="input-group-btn">
+							<button class="btn btn-lg btn-primary" type="button" onclick="getvalues()">Go!</button>
+						</span>
+					</div>
+				</form>
+			</div>
+
+		</div>
+		<!-- /row -->
+		<div class="row">
+			<div class="col-lg-12 text-center v-center">
+				<div id="result"></div>
+			</div>
+		</div>
+	</div>
+
+
+	<script src="http://ajax.googleapis.com/ajax/libs/jquery/2.0.2/jquery.min.js"></script>
+	<script src="assets/js/bootstrap.min.js"></script>
+	<script src="assets/js/regex.js"></script>
+
+	<script type="text/javascript">
+		$('#query-input').submit(function(ev) {
+			ev.preventDefault();
+			getvalues();
+			return false;
+		});
+	</script>
+
+</body>
+
 </html>
